Add build-tools install script
This commit contains a new tool in the tools/install-build-tools folder which installs all of the SEGA Genesis Runtime's build tools. These tools will be used to compile, assemble, link and perform other operations on both runtime library and user code. The script, install-build-tools.sh, is currently compatible with 32-bit and 64-bit Windows, 64-bit macOS and 64-bit Linux. It successfully builds the GCC toolchain for the Motorola 68000 on those operating systems. The toolchain is dynamically linked at the moment but may want to look into statically linking.