Modify install-gcc.sh to Zip amiga-gcc Toolchain
With this commit, there has been a shift in the way the build tools for sgr are stored. Previously there was
a 'build-tools' folder for storing the amiga-gcc toolchain for various platforms. However, in order to keep the
footprint of this repository light and to prevent users from having to use git lfs, the 'build-tools' folder and
all other auxiliary files for sgr will be stored in a seperate GitHub account acting as a CDN. To be specific,
there is an 'sgr-bucket' repository under the account, RRGamesCDN, which has GitHub Pages activated, allowing
resources to become available for download. Therefore, the install-gcc.sh script has been modified so as to not
place the m68k-amigaos-toolchain in this repository, but to zip the toolchain so it can be manually moved into
the 'sgr-bucket' repository under the RRGamesCDN account.

Example: wget https://rrgamescdn.github.io/sgr-bucket/build-tools/linux/64-bit/m68k-amigaos-toolchain.tar.xz