Add Preliminary Library Build System
This commit sees the addition of a build system for the SGR library
source code. It is implemented using the Make utility. Currently
it allows any number of subdirectories for source files and headers.
It is also generous with using spaces in filenames / filepaths. The
three key files for this commit are:

 - library-build.mk:     Calls library-rules.mk for each make goal
 - library-rules.mk:     Contains generated rules for compiling all
                         source files.
 - library-variables.mk: Contains important variables for building
                         the library code.

The makefiles for building an SGR project will be developed after
the library build system is complete. With that said, the current
state of the library build system is tested fairly well but by no
means considered stable. Later commits will seek to test the system
further and document the rather messy Make code.