Update Library Build System
The SGR library build system has been reinforced and now supports
spaces in file paths, even allowing SGR to be installed on a directory
containing spaces. It is able to find any .c and .asm 68000 sources
and compile them into the proper object files depeneding on build
config: release or debug. It can also auto clean the library, meaning
any object and dependency files that exist without a corresponding
source will be removed. It is not the quickest at the moment so
perhaps look into how it could be parallelized.