Add Basic Linker Script
Created a basic linker script that generates SEGA Genesis ROM images
from C and Assembly source files. Note that LTO seems to be an issue
at this point. GCC says it needs a plugin for LTO but amigaos gcc
isn't compatible. LTO has been tested in the past and works without
a plugin. This is something to look into in the future.