Change M68k Toolchain
This is a major commit changing the toolchain used to compile / assemble
68000 C and assembly files. Previously the m68k-amigaos target was used.
After several issues with the hunk file format such as relocation and
issues with the linker script, the target has been changed to m68k-elf.
The ELF file format is much more structured and is much better suited
for building ROM images. The makefiles have been restructured to use
this toolchain. install-gcc.sh under tools/install-build-tools/ has
been changed to build the new m68k-elf target. Note, the project currently
uses static libraries. This will be removed in the next commit to use
only SGR object files as the new toolchain has an issue with LTO archives.