Add Basic Support for VDP
Added (very) basic support for the VDP. Just reading and writing
registers as of now. There are functions both in C and hand optimized
assembly routines. The linker script was also modified slightly to
reorganize the ROM. Puts the VDP initial register values closer to
the boot code. Also, interrupt routines are at the end of ROM just
before read-only data.

You will notice a Doxyfile has been added. It was used to experiment
with project documentation. The final decision on documentation will
be made at a later date.