Add VDP Configuration and VDP I/O Functions
This commit adds the helpful C function VDPConfigure which takes in
a VDPConfig enum specifying what mode to put the VDP into and an
address to operate on. The VDP may be configured using this function
and data can be read from or written to it using VDPReadxxx and
VDPWritexxx functions accordingly (VDPReadWord, VDPWriteLongword, etc).

Note at the time of this commit, no attempt to optimize the C functions
has been made yet and hand optimimzed assembly not written.