Add VDP Register Functions
Some functions have been added for operating the VDP registers.
However, it is worth going back and accessing how functionality
will be implemented and in what language. Maintaining two versions
of the engine (C and ASM) will be cumbersome. Perhaps need to
implement small, inlinable functions in C and the rest in ASM.